Updated file.h for bugs and stupidity
It was a very stupid thing to try and intentionally mask
the char* with file as it will eventually lead to someone
forgetting to free(f) on a read file.  Also, I reduced some
error checking compelxity and eliminated an off by one byte
error in read_by_char that only manifested when running on the
actual corpus.