Stage 0.a Complete (Word Delimiter)
I have implemented the word delimiter extraction mechanism to work with
multiple character delimiters and single character delimiters.

It is supposed to throw an error if it finds a modular word delimiter
(meaning that there is one word delimiter for one situation and one
for another).

I still to need to optimize (removing garbage code and with valgrind),
but I did test on a small corpus with a two character delimiter and
also the real corpus with a one character delimiter.

...so, YAY!