Basic test code for remove garbage morpheme-candidates.
The code is code that currently makes assumptions and doesn't test
a few important things.  However, it does seem to work on prefixes
and suffixes that are simply appended (supplicative and other morpheme
types will be added soon).

This commit is happening now, so I can test on a larger file.